+++ src/main/java/com/utn/demo/services/IservicioGenerico.java
@@ -3,15 +3,15 @@ import java.util.List;
 
 public interface IservicioGenerico <E>{
 	
-	public E findById(int id) throws Exception;
+	public E findById(long id) throws Exception;
 	
 	public E save(E entityForm) throws Exception;
 	
-	public E update (int id, E entityForm) throws Exception;
+	public E update (long id, E entityForm) throws Exception;
 	
 	public List<E> findAll (int page, int size) throws Exception;
 	
-	public boolean delete (int id) throws Exception;
+	public boolean delete (long id) throws Exception;
 	
 	public int countPages(int size) throws Exception;
 }
